restructure dtds and selector functions
