+++ .travis.yml
@@ -13,7 +13,7 @@ matrix:
 
 install:
   - pip install scipy
-  - pip install .  # Test installation correctness
+  - pip install .       # Test installation correctness
   - pip install .[dev]  # Test development dependencies
 
 script:
